pyproject.toml
[project]
name = "bitterroot"
version = "0.1.0"
description = "A hostable, searchable, database to categorize and analyze the scholarships of different universities and identify different niches of academic literature."
authors = [
    { name = "Noah S. Roberts", email = "35052448+Mase3206@users.noreply.github.com" },
    { name = "John Norton", email = "unknown@example.com" },
]
requires-python = ">3.13,<3.14"
readme = "README.md"
dependencies = [
    "django (>=5.2.7,<6.0.0)",
    "psycopg[binary] (>=3.2.12,<4.0.0)",
    "django-debug-toolbar (>=6.1.0,<7.0.0)",
    "django-rest-framework (>=0.1.0,<0.2.0)",
    "django-extensions (>=4.1,<5.0)",
    "environs (>=14.5.0,<15.0.0)",
    "django-cors-headers (>=4.9.0,<5.0.0)",
    "whitenoise>=6.11.0",
    "gunicorn>=23.0.0",
    "django-allauth[headless]>=65.13.1",
    "requests>=2.32.5",
    "django-modern-csrf>=1.0.1",
]

[project.scripts]
manage = "bitterroot.manage:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "inflection>=0.5.1",
    "pyyaml>=6.0.3",
    "ruff (>=0.14.3,<0.15.0)",
    "uritemplate>=4.2.0",
]


# Ruff settings

# [tool.ruff]
# include = ["bitterroot/**/*.py"]

[tool.ruff.format]
docstring-code-format = true
indent-style = "space"
line-ending = "lf"
skip-magic-trailing-comma = false

[tool.ruff.lint]
select = [
    "W291", "F401", "I"
]
[tool.ruff.lint.isort]
combine-as-imports = true
